# openapi.yaml
openapi: 3.0.0
info:
  title: order-api
  version: 1.0.0
# servers:
#   - url: http://orderservice-3556955799:9090
paths:
  /order:
    post:
      tags:
        - default
      summary: create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}

    get:
      tags:
        - default
      summary: get
      responses:
        "200":
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
      deprecated: false
      security: []
  /order/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    put:
      tags:
        - default
      summary: update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}

    get:
      tags:
        - default
      summary: get by ID
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderInput"

    delete:
      tags:
        - default
      summary: delete
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}

components:
  schemas:
    OrderInput:
      type: object
      properties:
        ID:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              ItemID:
                type: string
              Quantity:
                type: integer
        Total:
          type: number